Compiles on linux.

+++ src/light/infinite.cpp
@@ -5,6 +5,7 @@
 #include "integrator/surface.hpp"
 #include "scene/scene.hpp"
 #include "material/material.hpp"
+#include <complex>
 
 namespace _462 {
 
@@ -16,7 +17,7 @@ InfiniteAreaLight::InfiniteAreaLight(const Matrix3 &l2w, Material *m_ptr)
 	float *tex_buffer = new float[width * height];
 
 	for (uint32_t y = 0; y < height; y++) {
-		float sin_theta = std::sinf(PI * (y + 0.5f) / height);
+		float sin_theta = std::sin(PI * (y + 0.5f) / height);
 		float inv_y = height - 1 - y;
 		for (uint32_t x = 0; x < width; x++) {
 			tex_buffer[y * width + x] = m_ptr->get_filtered_texture_pixel(x, inv_y).relative_luminance();
@@ -37,8 +38,8 @@ Color3 InfiniteAreaLight::sample_L(const Vector3 &p, //float pEpsilon,
 	float uv_pdf;
 	Vector2 uv = dist_ptr->sample_continuous(r1, r2, &uv_pdf);
 	float theta = PI * uv.y, phi = 2 * PI *	uv.x;
-	float cos_theta = std::cosf(theta), sin_theta = std::sinf(theta);
-	float sin_phi = std::sinf(phi), cos_phi = std::cosf(phi);
+	float cos_theta = std::cos(theta), sin_theta = std::sin(theta);
+	float sin_phi = std::sin(phi), cos_phi = std::cos(phi);
 
 	Vector3 shape_p(sin_theta * cos_phi, sin_theta * sin_phi, cos_theta);
 	*wi = LightToWorld * shape_p;
@@ -61,9 +62,9 @@ Color3 InfiniteAreaLight::Power(const Scene *scene) const {
 
 float InfiniteAreaLight::pdf(const Vector3 &p, const Vector3 &wi) const {
 	Vector3 trans_wi = WorldToLight * wi;
-	float theta = std::acosf(trans_wi.z);
-	float sin_theta = std::sinf(theta);
-	float phi = std::atan2f(trans_wi.y, trans_wi.x);
+	float theta = std::acos(trans_wi.z);
+	float sin_theta = std::sin(theta);
+	float phi = std::atan2(trans_wi.y, trans_wi.x);
 	phi = (phi < 1e-5f) ? phi + 2 * PI : phi;
 
 	float u = phi / (2 * PI), v = theta / PI;
@@ -72,8 +73,8 @@ float InfiniteAreaLight::pdf(const Vector3 &p, const Vector3 &wi) const {
 
 Color3 InfiniteAreaLight::Le(const Ray &r) const {
 	Vector3 trans_d = WorldToLight * r.d;
-	float theta = std::acosf(trans_d.z);
-	float phi = std::atan2f(trans_d.y, trans_d.x);
+	float theta = std::acos(trans_d.z);
+	float phi = std::atan2(trans_d.y, trans_d.x);
 	phi = (phi < 1e-5f) ? phi + 2 * PI : phi;
 
 	float u = phi / (2 * PI), v = theta / PI;
